+++ src/modules/test/unitTests/main.c
@@ -1,5 +1,5 @@
 #include "../testdll.h"
-#include <stdbool.h>
+#include "../../../unitTestLib/unitTest.h"
 
 #define LENGTH 5
 
@@ -72,20 +72,7 @@ bool testIfCreatedInterfacesAreValid(){
     return true;
 }
 
-void iterateTests(char* names[], bool (*funcs[])(), uint16_t length)
-{
-    uint16_t fails = 0;
-    for (size_t i = 0; i < length; ++i){
-        printf("%s: ", names[i]);
-        if (funcs[i]()){
-            printf("PASS\n");
-            continue;
-        }
-        ++fails;
-        printf("FAIL\n");
-    }
-    printf("\n%u tests ran, %u failed", length, fails);
-}
+
 
 int main()
 {
