+++ src/unitTestLib/unitTest.h
@@ -1,0 +1,10 @@
+#ifndef UNIT_TEST_LIB
+#define UNIT_TEST_LIB
+
+#include <stdint.h>
+#include <stdbool.h>
+#include <stdio.h>
+
+extern void iterateTests(char* [], bool (*[])(), uint16_t);
+
+#endif
